requirements.txt
pillow 
scikit-image 
numpy
tensorflow 
scikit-learn
